fix: update getWorkspacesByRepo function signature to be more consistent with base function
